Add metrics query
